backend/requirement.txt
Flask
flask-cors
pymupdf
python-docx
ollama
